+++ docs/src/pages/components/searchfield/examples/index.ts
@@ -1,0 +1,6 @@
+export { DefaultSearchFieldExample } from './DefaultSearchFieldExample';
+export { PlaceholderSearchFieldExample } from './PlaceholderSearchFieldExample';
+export { RefExample } from './refs';
+export { SearchFieldStyledPropsExample } from './SearchFieldStyledPropsExample';
+export { SizeSearchFieldExample } from './SizeSearchFieldExample';
+export { VariationSearchFieldExample } from './VariationSearchFieldExample';