+++ examples/next/pages/ui/components/authenticator/sign-in-sms-mfa/index.page.tsx
@@ -1,10 +1,18 @@
-import { Amplify } from 'aws-amplify';
+import { Amplify, I18n } from 'aws-amplify';
 
 import { Authenticator } from '@aws-amplify/ui-react';
 import '@aws-amplify/ui-react/styles.css';
 
+import { translations } from '@aws-amplify/ui';
 import awsExports from './aws-exports';
 Amplify.configure(awsExports);
+
+I18n.putVocabularies(translations);
+I18n.setLanguage('en');
+
+I18n.putVocabulariesForLanguage('en', {
+  'Invalid code or auth state for the user.': 'translated text',
+});
 
 export default function AuthenticatorWithSmsMfa() {
   return (