+++ packages/react/src/components/Authenticator/Router/index.tsx
@@ -1,11 +1,11 @@
 import { CognitoUserAmplify } from '@aws-amplify/ui';
-import * as React from 'react';
 
 import { useAuthenticator } from '..';
 import { View } from '../../..';
 import { ConfirmSignIn } from '../ConfirmSignIn';
 import { ConfirmSignUp } from '../ConfirmSignUp';
 import { ForceNewPassword } from '../ForceNewPassword';
+import { useCustomComponents } from '../hooks/useCustomComponents';
 import { ConfirmResetPassword, ResetPassword } from '../ResetPassword';
 import { SetupTOTP } from '../SetupTOTP';
 import { SignInSignUpTabs } from '../shared';
@@ -13,7 +13,7 @@ import { ConfirmVerifyUser, VerifyUser } from '../VerifyUser';
 
 export type RouterProps = {
   className?: string;
-  children: ({
+  children?: ({
     signOut,
     user,
   }: {
@@ -32,15 +32,15 @@ export function Router({
   className,
   variation = 'default',
 }: RouterProps) {
+  const { route, signOut, user } = useAuthenticator();
+
   const {
     components: { Header, Footer },
-    route,
-    signOut,
-    user,
-  } = useAuthenticator();
+  } = useCustomComponents();
 
+  // `Authenticator` might not have `children` for non SPA use cases.
   if (['authenticated', 'signOut'].includes(route)) {
-    return children({ signOut, user });
+    return children ? children({ signOut, user }) : null;
   }
 
   return (