+++ packages/vue/src/components/authenticator.vue
@@ -1,6 +1,14 @@
 <script setup lang="ts">
 import { useAuth } from '../composables/useAuth';
-import { ref, computed, useAttrs, watch, Ref, onMounted } from 'vue';
+import {
+  ref,
+  computed,
+  useAttrs,
+  watch,
+  Ref,
+  onMounted,
+  onUnmounted,
+} from 'vue';
 import { useActor, useInterpret } from '@xstate/vue';
 import {
   getActorState,
@@ -10,6 +18,7 @@ import {
   translate,
   CognitoUserAmplify,
   SocialProvider,
+  listenToAuthHub,
 } from '@aws-amplify/ui';
 
 import SignIn from './sign-in.vue';
@@ -61,11 +70,13 @@ const emit = defineEmits([
 const machine = createAuthenticatorMachine();
 
 const service = useInterpret(machine);
+let unsubscribeHub: ReturnType<typeof listenToAuthHub>;
 
 const { state, send } = useActor(service);
 useAuth(service);
 
 onMounted(() => {
+  unsubscribeHub = listenToAuthHub(send);
   send({
     type: 'INIT',
     data: {
@@ -76,6 +87,10 @@ onMounted(() => {
       services,
     },
   });
+});
+
+onUnmounted(() => {
+  if (unsubscribeHub) unsubscribeHub();
 });
 
 const actorState = computed(() => getActorState(state.value));
@@ -401,6 +416,9 @@ const hasTabs = computed(() => {
           </template>
           <template #header>
             <slot name="force-new-password-header"></slot>
+          </template>
+          <template #force-new-password-form-fields>
+            <slot name="force-new-password-form-fields"></slot>
           </template>
           <template
             #footer="{ onHaveAccountClicked, onForceNewPasswordSubmit }"